requirements.txt
llama-cpp-python==0.2.90
gradio==4.44.0
huggingface_hub